Rename bqutil -> bqext
