+++ bqext/README.md
@@ -1,0 +1,6 @@
+# The bqext package
+
+This package contains extensions for the bigquery library, to facilitate
+various common operations, notably query processing.
+
+
