--- bqutil/README.md
@@ -1,6 +1,0 @@
-# The bqutil package
-
-This package contains utilities that facilitate various bigquery
-operations, notably query processing.
-
-
